@@ -13,9 +13,9 @@
     "jquery": "^3.5.1",
     "prop-types": "^15.7.2",
     "rc-steps": "^4.1.3",
-    "rc-tooltip": "^5.0.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0-rc.2",
+    "react-bootstrap": "^2.2.1",
+    "react-dom": "^18.0.0-rc.2",
     "react-final-form": "^6.5.2",
     "react-redux": "^7.2.2",
     "react-router-dom": "^6.2.1",
